+++ plot/viewer/src/canvases/ServiceDetailCanvas.tsx
@@ -17,7 +17,6 @@ export function ServiceDetailCanvas(props: SketchCanvasProps) {
       hideRootServiceNode={true}
       shouldDrill={shouldDrillActorRef}
       showFoldButton={true}
-      injectAnchor={false}
     />
   );
 }